# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Free mentors is web application where Mentees can request mentorship sessions to Mentors, Users can signup and singin to web application, Admin can change user to mentor, Users can view mentors and view a specific mentor, Mentors can accept or reject mentorship request from mentees, Mentees can review Mentors after mentorship, Admin can delete review.",
    "version": "1.0.0",
    "title": "Free Mentors Web Application",
    "termsOfService": "https://github.com/erickyvand",
    "contact": {
      "email": "erickyvand@gmail.com"
    }
  },
  "host": "the-freementors.herokuapp.com",
  "basePath": "/api/v2",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/sessions/{sessionId}/review": {
      "delete": {
        "tags": [
          ""
        ],
        "summary": "Admin delete a r",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "in": "path",
            "name": "sessionId",
            "description": "Delete a review"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "The review ID not found"
          }
        }
      }
    },
...
    "/user/{userId}": {
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "Admin change user to mentor",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "in": "path",
            "name": "userId",
            "description": "Change User to mentor",
            "required": true
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User Id not found"
          },
          "409": {
            "description": "User already changed to mentor"
          },
          "200": {
            "description": "User account changed to mentor"
          }
        }
      }
    },
...
    "/mentors/{mentorId}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "View a specific mentor",
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "in": "path",
            "name": "mentorId",
            "description": "View a specific mentor"
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Mentor with the given ID does not exists"
          },
          "400": {
            "description": "The Given ID is not for a mentor"
          },
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
...
}